Adding ejection action and operation
While writing the tests for the ejection action, I noticed
a bunch of spilling over from test to test. The split here
should allow for easier testing for the relevant user actions.
